Allows for asynchronous kernel execution
At least when executing via the V3D register-poking,
we now execute the kernel and do not wait immediately for the execution
to finish, but instead prepare the UNIFORMs for the next iteration first.

For that to work, the UNIFORM buffer is duplicated, so we can prepare one
while the other is being used. After every execution, the buffers are switched.

Improves full-program performance by some 1-5% for the hand full of
test-cases run, but should have greater impact on longer-running kernels.